Add TempResponseCart
This type will be used instead of a ResponseCart to internally keep the
cart state. It is identical to ResponseCart except that it can contain
products that have incomplete data (RequestCartProducts). These products
are filtered out before returning the cart for use, so the API of the
package remains the same.

This allows us to clearly define what type of cart we're working with in
any given location, and removes the ambiguity of the
TempResponseCartProduct that previously had to be dealt with everywhere.
Now we can guarantee that the products in a ResponseCart will always
have the data returned by the server. Any new products added that have
not yet made the round-trip to the server will be hidden from consumers
until the server sends them back. In the meantime, consumers should
respect the `isLoading` and `isPendingUpdate` flags that indicate that
the cart data is currently not up-to-date.