Modify usePrepareProductsForCart to handle plans like products
Previously, there were separate mechanisms in usePrepareProductsForCart
for plans and other products. However, this was unnecessary for several
reasons.

Firstly, plans are not that different from other products; they have a
slug and an ID. Secondly, the slug may need to be translated from an
alias, but that's also true of other products; the main difference is
that plan slugs are aliased to a "path slug" which is stored in a
hard-coded list in lib/plans. Since they are not fetched from a server,
we don't need to request them or wait for them.

This also replaces the various product creation functions with
createRequestCartProduct, which can handle all of them. There were some
special cases being handled using specific product creation functions,
but if you examine those functions, they were generic, so we're able to
remove those special cases and the conditions that triggered them (eg:
Jetpack products).